serious bugs fixs and support multi layers GRU 1. fixed masked prob calculation need to sum -1 dim and keepdim 2. normalized returns
discovered issues:
1. gru will becomes all -1 and 1. policy result and values will becomes the same on different stages.
2. returns need to be normaliized. the network will become returning nan if the loss is too large. (still checking if it cause the issue 1)

+++ games/puzzle2048.py
@@ -44,7 +44,7 @@ class Puzzle2048(Game):
         self.game = Src(self.size)
 
     def getState(self, playerId: int):
-        state = np.zeros(self.observationShape, dtype=int)
+        state = np.zeros(self.observationShape, dtype=float)
         # state = np.zeros((1, self.size, self.size), dtype=int)
         for _, _, cell in self.game.grid.eachCell():
             if cell:
